// src/api/models/RoomInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RoomInput = {
    /**
     * Room label
     */
    label: string;
    /**
     * Room maximum capacity (number of people)
     */
    max_size: number;
};